refactoring of CIVL refinement code
