+++ Source/Concurrency/CivlTypeChecker.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using Microsoft.Boogie;
 using System.Diagnostics.Contracts;
 using System.Diagnostics;
 using Microsoft.Boogie.GraphUtil;
