+++ Source/Concurrency/Concurrency.csproj
@@ -74,14 +74,18 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="YieldingProcInstrumentation.cs" />
     <Compile Include="LinearTypeChecker.cs" />
     <Compile Include="MoverCheck.cs" />
-    <Compile Include="CivlRefinement.cs" />
+    <Compile Include="YieldingProcChecker.cs" />
     <Compile Include="CivlVCGeneration.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="RefinementInstrumentation.cs" />
     <Compile Include="SimulationRelation.cs" />
     <Compile Include="CivlTypeChecker.cs" />
     <Compile Include="TransitionRelationComputation.cs" />
+    <Compile Include="NoninterferenceInstrumentation.cs" />
+    <Compile Include="YieldingProcDuplicator.cs" />
     <Compile Include="YieldTypeChecker.cs" />
     <Compile Include="CivlUtil.cs" />
   </ItemGroup>
