+++ Source/Concurrency/MoverCheck.cs
@@ -1,9 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Diagnostics.Contracts;
-using System.Diagnostics;
 
 namespace Microsoft.Boogie
 {
