+++ Source/Concurrency/YieldTypeChecker.cs
@@ -1,11 +1,7 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using Microsoft.Boogie;
-using System.Diagnostics.Contracts;
-using Microsoft.Boogie.AbstractInterpretation;
 using Microsoft.Boogie.GraphUtil;
 using System.Diagnostics;
 using System.ComponentModel;
